refactor: update component imports to new admin directory structure and add global styles
